Fix serif font on systems that do not define a system font
Some linux distributions that do not use a desktop environment do
not define a system font. On these systems, the system-font css file
does not work if none of the font families are matched (and it
seems chrome ignores it anyway).

I added sans-serif as a fallback font, which is usually always defined
in fontconfig as a sans-serif font.

Without this, cerebro uses an ugly ans obscure serif font due to
no fallback having been found.