feat(main window): display main window on the screen containing or closest to the mouse pointer
When displaying the main window for Cerebro so far the primary screen has been chosen. This is
inconvenient for workstations with more than 1 screen. After this commit is applied the main window
will be displayed in the screen containing the mouse pointer (or the screen closest to the pointer
if the pointer is in between screens).

fix cerebroapp/cerebro#647